requirements.txt
torch==1.11.0
torchvision==0.12.0
timm>=0.5.4
numpy>=1.21.4
Pillow>=9.1.0
matplotlib>=3.2.2
flask>=1.1.2
opencv-python
tqdm
tabulate